test implementation attempt...

+++ WebTrader/Classes/Order.cs
@@ -10,7 +10,7 @@ namespace Trader.objects
         public DateTime Time { get; set; }
         public string Type { get; set; } // buy/sell
         public string Kind { get; set; } // limit/;
-        public float Amount { get; set; }
-        public float Price { get; set; }
+        public double Amount { get; set; }
+        public double Price { get; set; }
     }
 }
